+++ game/shared/shareddefs.h
@@ -14,7 +14,7 @@
 #define TICK_INTERVAL			(gpGlobals->interval_per_tick)
 
 
-#define TIME_TO_TICKS( dt )		( (int)( 0.5f + (float)(dt) / TICK_INTERVAL ) )
+#define TIME_TO_TICKS( dt )		( (int)( 0.5f + (double)(dt) / TICK_INTERVAL ) )
 #define TICKS_TO_TIME( t )		( TICK_INTERVAL *( t ) )
 #define ROUND_TO_TICKS( t )		( TICK_INTERVAL * TIME_TO_TICKS( t ) )
 #define TICK_NEVER_THINK		(-1)
